# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TownRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "A complete SimConfig (wins over preset/overrides).",
            "title": "Config"
          },
          "houses": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 20.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Houses"
          },
          "overrides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Deep-merged SimConfig overrides.",
            "title": "Overrides"
          },
          "preset": {
            "default": "village",
            "description": "Base preset.",
            "title": "Preset",
            "type": "string"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "seed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Master seed override.",
            "title": "Seed"
          }
        },
        "title": "TownRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Seeded utility-town engine: geography, networks, customers, simulation, meter-to-cash.",
    "title": "utilsim",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config/presets/{name}": {
      "get": {
        "operationId": "get_preset_api_config_presets__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preset"
      }
    },
...
    "/api/towns": {
      "post": {
        "description": "Generate a town from a preset (with seed, houses, scenario and deep-merged overrides) or a complete\n``config``. ``{townId, status}``: 201 ready, 202 building (poll ``GET /api/towns/{id}``), 422 invalid or failed.",
        "operationId": "create_town_api_towns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TownRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Town"
      }
    },
    "/api/towns/{tid}": {
      "get": {
        "operationId": "get_town_api_towns__tid__get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Town"
      }
    },
...
    "/api/towns/{tid}/snapshot.json": {
      "get": {
        "description": "The town's ``utility-town/2.0`` snapshot (gzip); ``profile`` (or ``detail``) ``viewer`` leaves out reads and\nthe customer and billing tables.",
        "operationId": "get_snapshot_api_towns__tid__snapshot_json_get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "title": "Tid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "full",
                    "viewer"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile"
            }
          },
          {
            "description": "Alias of profile.",
            "in": "query",
            "name": "detail",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "full",
                    "viewer"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Alias of profile.",
              "title": "Detail"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Snapshot"
      }
    },
...
  }
}